Cleanup unused plot routes
Cleanup subsequently redundant components.
Point `/scat-cs/[id]` to `/scat-cs/inspect?ids={id}`
