--- app/src/ScatteringCrossSectionSet/HistoryTable.module.css
@@ -1,13 +1,0 @@
-/*
- * SPDX-FileCopyrightText: LXCat team
- *
- * SPDX-License-Identifier: AGPL-3.0-or-later
- */
-
-.htable {
-  width: 100%;
-}
-
-.htable th {
-  text-align: left;
-}
